Retrieve AlphaMissense Resources as DuckDB Databases

ALPHAMISSENSE_RECORD is a constant identifier corresponding to the default version of the AlphaMissense resource to use.

am_browse() opens a web browser at the Zenodo record for the AlphaMissense data.

am_available() reports available datasets in the record.

am_data() retrieves a single key from the the AlpahMissense Zenodo site and parses the file into a DuckDB database.

Usage

ALPHAMISSENSE_RECORD

am_browse(record = ALPHAMISSENSE_RECORD)

am_available(record = ALPHAMISSENSE_RECORD, bfc = BiocFileCache())

am_data(
  key,
  record = ALPHAMISSENSE_RECORD,
  bfc = BiocFileCache(),
  as = c("tbl", "tsv")
)

Format

An object of class character of length 1.

Arguments

record

character(1) Zenodo record for the AlphaMissense data resources.

bfc

an object returned by BiocFileCache() representing the location where downloaded files and the parsed database will be stored. The default is the 'global' BiocFileCache.

key

a character(1) 'key' from the result of am_available(), or a single row of the tibble returned by am_available().

as

chracter(1) type of return value.

"tbl": a dbplyr tbl representation of the database resource.

"tsv": path to the tsv.gz file representing the resource and downloaded from Zenodo

Value

am_available() returns a tibble with columns key, size, and link. The meaning of key must be determined with reference to the information at am_browse().

am_data() returns a dbplyr (database) tibble represented the downloaded and parsed file. Fields in the database are as described on the Zenodo resource page.

Details

ALPHAMISSENSE_RECORD can be set before the package is loaded with the environment variable of the same name, e.g., Sys.setenv(ALPHAMISSENSE_RECORD = "8208688"). The default is the most recent version (version 2) on 25 September, 2023.

am_data() uses BiocFileCache to download and store the file and the corresponding DuckDB database.

Examples

ALPHAMISSENSE_RECORD
#> [1] "8360242"

if (interactive())
    am_browse()

am_available()
#> # A tibble: 7 × 5
#>   record  key                             size cached link                      
#>   <chr>   <chr>                          <dbl> <lgl>  <chr>                     
#> 1 8360242 aa_substitutions          1207278510 TRUE   https://zenodo.org/api/fi…
#> 2 8360242 gene_hg19                     243943 TRUE   https://zenodo.org/api/fi…
#> 3 8360242 gene_hg38                     253636 TRUE   https://zenodo.org/api/fi…
#> 4 8360242 hg19                       622293310 TRUE   https://zenodo.org/api/fi…
#> 5 8360242 hg38                       642961469 TRUE   https://zenodo.org/api/fi…
#> 6 8360242 isoforms_aa_substitutions 2461351945 TRUE   https://zenodo.org/api/fi…
#> 7 8360242 isoforms_hg38             1177361934 TRUE   https://zenodo.org/api/fi…

am_data("hg38")
#> # Source:   table<hg38> [?? x 10]
#> # Database: DuckDB 0.8.1 [root@Darwin 21.6.0:R 4.3.1//Users/ma38727/Library/Caches/org.R-project.R/R/BiocFileCache/dbb12417e9ac_dbb12417e9ac]
#>    `#CHROM`   POS REF   ALT   genome uniprot_id transcript_id    protein_variant
#>    <chr>    <dbl> <chr> <chr> <chr>  <chr>      <chr>            <chr>          
#>  1 chr1     69094 G     T     hg38   Q8NH21     ENST00000335137… V2L            
#>  2 chr1     69094 G     C     hg38   Q8NH21     ENST00000335137… V2L            
#>  3 chr1     69094 G     A     hg38   Q8NH21     ENST00000335137… V2M            
#>  4 chr1     69095 T     C     hg38   Q8NH21     ENST00000335137… V2A            
#>  5 chr1     69095 T     A     hg38   Q8NH21     ENST00000335137… V2E            
#>  6 chr1     69095 T     G     hg38   Q8NH21     ENST00000335137… V2G            
#>  7 chr1     69097 A     G     hg38   Q8NH21     ENST00000335137… T3A            
#>  8 chr1     69097 A     C     hg38   Q8NH21     ENST00000335137… T3P            
#>  9 chr1     69097 A     T     hg38   Q8NH21     ENST00000335137… T3S            
#> 10 chr1     69098 C     A     hg38   Q8NH21     ENST00000335137… T3N            
#> # ℹ more rows
#> # ℹ 2 more variables: am_pathogenicity <dbl>, am_class <chr>

## close the connection opened when adding the data
db_disconnect()