Manipulate the Database of Missense Mutations

db_tables() queries for the names of temporary and regular tables defined in the database.

db_temporary_table() creates a temporary (for the duration of the duckdb connection) table from a tibble.

db_range_join() performs a range join, finding all positions in key within ranges defined by join. The result is stored in table to.

db_disconnect() disconnects the duckdb database and shuts down the DuckDB server associated with the connection. Temporary tables are lost.

Usage

db_tables(db)

db_temporary_table(db, value, to)

db_range_join(db, key, join, to)

db_disconnect(db)

Arguments

db

a duckdb_connection object, as returned by am_data().

value

a data.frame / tibble containing data to be placed in a temporary table, e.g., from a GenomicRanges object to be used in a range join.

to

the character(1) name of the table to be created

key

a character(1) table name in db containing missense mutation coordinates.

join

a character(1) table name in db containing ranges to be used for joining with (filtering) key.

Value

db_tables() returns a character vector of database table names.

db_temporary_table() returns the temporary table as a dbplyr tibble.

db_range_join() returns to (the temporary table created from the join) as a dbplyr tibble.

db_disconnect() returns FALSE if the connection has already been closed or is not valid (via dbIsValid()) or TRUE if disconnection is successful. Values are returned invisibly.

Details

db_temporary_table() overwrites an existing table with name to.

db_range_join() overwites and existing table with name to.

The table key is usually "hg19" or "hg38" and must have #CHROM and POS columns. The table join must have columns #CHROM, start and end. Following Bioconductor convention and as reported in at am_browse(), coordinates are 1-based and ranges defined by start and end are closed. All columns from both key and join are included, so column names (other than #CHROM) cannot be duplicated.

Examples

db <- am_data("hg38")
db_tables(db)
#> [1] "gene_hg19" "gene_hg38" "hg19"      "hg38"     

## ranges of interest -- the first 200000 bases on chromsomes 1-4.
ranges <- tibble(
    `#CHROM` = paste0("chr", 1:4),
    start = rep(1, 4),
    end = rep(200000, 4)
)
db_temporary_table(db, ranges, "ranges")
#> # Source:   table<ranges> [4 x 3]
#> # Database: DuckDB 0.8.1 [ma38727@Darwin 21.6.0:R 4.3.1//Users/ma38727/Library/Caches/org.R-project.R/R/BiocFileCache/dbb12417e9ac_dbb12417e9ac]
#>   `#CHROM` start    end
#>   <chr>    <dbl>  <dbl>
#> 1 chr1         1 200000
#> 2 chr2         1 200000
#> 3 chr3         1 200000
#> 4 chr4         1 200000

db_tables(db)
#> [1] "gene_hg19" "gene_hg38" "hg19"      "hg38"      "ranges"   

rng <- db_range_join(db, "hg38", "ranges", "ranges_overlaps")
#> * [13:40:33][info] doing range join of 'hg38' with 'ranges'
rng
#> # Source:   table<ranges_overlaps> [?? x 12]
#> # Database: DuckDB 0.8.1 [ma38727@Darwin 21.6.0:R 4.3.1//Users/ma38727/Library/Caches/org.R-project.R/R/BiocFileCache/dbb12417e9ac_dbb12417e9ac]
#>    `#CHROM`   POS REF   ALT   genome uniprot_id transcript_id    protein_variant
#>    <chr>    <dbl> <chr> <chr> <chr>  <chr>      <chr>            <chr>          
#>  1 chr2     45769 G     T     hg38   Q1W6H9     ENST00000327669… S206Y          
#>  2 chr2     45770 A     C     hg38   Q1W6H9     ENST00000327669… S206A          
#>  3 chr2     45770 A     G     hg38   Q1W6H9     ENST00000327669… S206P          
#>  4 chr2     45770 A     T     hg38   Q1W6H9     ENST00000327669… S206T          
#>  5 chr2     45771 G     T     hg38   Q1W6H9     ENST00000327669… S205R          
#>  6 chr2     45771 G     C     hg38   Q1W6H9     ENST00000327669… S205R          
#>  7 chr2     45772 C     T     hg38   Q1W6H9     ENST00000327669… S205N          
#>  8 chr2     45772 C     A     hg38   Q1W6H9     ENST00000327669… S205I          
#>  9 chr2     45772 C     G     hg38   Q1W6H9     ENST00000327669… S205T          
#> 10 chr2     45773 T     G     hg38   Q1W6H9     ENST00000327669… S205R          
#> # ℹ more rows
#> # ℹ 4 more variables: am_pathogenicity <dbl>, am_class <chr>, start <dbl>,
#> #   end <dbl>
rng |>
    count(`#CHROM`) |>
    arrange(`#CHROM`)
#> # Source:     SQL [3 x 2]
#> # Database:   DuckDB 0.8.1 [ma38727@Darwin 21.6.0:R 4.3.1//Users/ma38727/Library/Caches/org.R-project.R/R/BiocFileCache/dbb12417e9ac_dbb12417e9ac]
#> # Ordered by: `#CHROM`
#>   `#CHROM`     n
#>   <chr>    <dbl>
#> 1 chr1      2018
#> 2 chr2      2028
#> 3 chr4      7488

db_disconnect(db)